feat(db): check for existing play by parent hash
